Extract auth API endpoints
