+++ app/lib/api/apiWrapper.ts
@@ -1,11 +1,13 @@
 import * as maps from './requests/maps';
 import * as users from './requests/users';
 import * as replays from './requests/replays';
+import * as auth from './requests/auth';
 
 const api = {
     maps,
     users,
     replays,
+    auth,
 };
 
 export default api;
