+++ app/lib/api/requests/auth.ts
@@ -1,32 +1,10 @@
-import axios from 'axios';
-import apiInstance from './apiInstance';
-
-const getRedirectUri = () => {
-    if (typeof window === 'undefined') {
-        // Avoid some nextjs compilation errors regarding window being undefined
-        return undefined;
-    }
-    return `${window.location.origin}/auth_redirect`;
-};
-
-export const generateAuthUrl = (state: string): string => {
-    const url = 'https://api.trackmania.com/oauth/authorize';
-
-    const params = {
-        response_type: 'code',
-        client_id: process.env.NEXT_PUBLIC_CLIENT_ID,
-        redirect_uri: getRedirectUri(),
-        state,
-    };
-
-    return axios.getUri({ url, params });
-};
+import { getRedirectUri } from '../../utils/auth';
+import apiInstance from '../apiInstance';
 
 interface AuthorizationResponse {
     displayName: string;
     accountId: string;
 }
-
 export const authorizeWithAccessCode = async (
     accessCode: string, clientCode?: string,
 ): Promise<AuthorizationResponse> => {
