+++ app/lib/utils/auth.ts
@@ -1,0 +1,22 @@
+import axios from 'axios';
+
+export const getRedirectUri = () => {
+    if (typeof window === 'undefined') {
+        // Avoid some nextjs compilation errors regarding window being undefined
+        return undefined;
+    }
+    return `${window.location.origin}/auth_redirect`;
+};
+
+export const generateAuthUrl = (state: string): string => {
+    const url = 'https://api.trackmania.com/oauth/authorize';
+
+    const params = {
+        response_type: 'code',
+        client_id: process.env.NEXT_PUBLIC_CLIENT_ID,
+        redirect_uri: getRedirectUri(),
+        state,
+    };
+
+    return axios.getUri({ url, params });
+};
